+++ gala/potential/__init__.py
@@ -2,3 +2,4 @@ from .potential import *
 from .hamiltonian import *
 from .frame import *
 from .scf import SCFPotential
+from .common import PotentialParameter